# swagger_docs.yaml
openapi: 3.0.2
info:
  title: StoryBoom AI API
  description: |
    API for StoryBoom AI - Case Study Generation Platform
    
    This API provides endpoints for:
    - User authentication and management
    - Case study creation and management
    - Interview management (provider and client)
    - Media generation (videos, podcasts)
    - Feedback collection and analysis
    - Metadata analysis
    
    ## Authentication
    Most endpoints require authentication. Use session-based authentication by logging in first.
    
    ## Rate Limits
    API calls are limited to prevent abuse. Contact support if you need higher limits.
    
    ## IP Access
    For production access, contact support to whitelist your IP address.
  version: 1.0.0
  contact:
    name: StoryBoom AI Support
    email: storyboomai@gmail.com
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:10000
    description: Development server
  - url: https://your-production-domain.com
    description: Production server
security:
  - ApiKeyAuth: []
components:
  securitySchemes:
    ApiKeyAuth:
      type: apiKey
      in: header
      name: Authorization
      description: Session-based authentication (login required)
...
paths:
...
  /api/save_provider_summary:
    post:
      tags:
        - Interviews
      summary: Save provider interview summary
      description: Save the summary from a solution provider interview
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - case_study_id
                - summary
              properties:
                case_study_id:
                  type: integer
                summary:
                  type: string
      responses:
        '200':
          description: Provider summary saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  case_study_id:
                    type: integer
                  status:
                    type: string
  /api/save_client_transcript:
    post:
      tags:
        - Interviews
      summary: Save client interview transcript
      description: Save transcript from a client interview (requires token)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - transcript
                - token
              properties:
                transcript:
                  type: string
                token:
                  type: string
      responses:
        '200':
          description: Client transcript saved successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                  filename:
                    type: string
  /api/generate_client_summary:
    post:
      tags:
        - Interviews
      summary: Generate client summary
      description: Generate summary from client interview transcript
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - transcript
                - token
              properties:
                transcript:
                  type: string
                token:
                  type: string
      responses:
        '200':
          description: Client summary generated successfully
          content:
            application/json:
              schema:
                type: object
                properties:
                  client_summary:
                    type: string
                  status:
                    type: string
  
  # Media generation endpoints
...